backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary
python-dotenv
dj_database_url
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pymysql
google-auth
requests